Added ability to update grocerylist properties

+++ Backend/Application/GroceryListService.cs
@@ -49,8 +49,23 @@ public class GroceryListService : IGroceryListService
         return _groceryListRepository.All();
     }
 
-    public GroceryList DeleteList(int id)
+    public GroceryList DeleteList(int id, GroceryList groceryList)
     {
+        if (id != groceryList.Id)
+            throw new ValidationException("ID in body does not match route.");
+        var validation = _validator.Validate(groceryList);
+        if (!validation.IsValid)
+            throw new ValidationException(validation.ToString());
         return _groceryListRepository.Delete(id);
+    }
+
+    public GroceryList UpdateList(int id, GroceryList groceryList)
+    {
+        if (id != groceryList.Id)
+            throw new ValidationException("ID in body does not match route.");
+        var validation = _validator.Validate(groceryList);
+        if (!validation.IsValid)
+            throw new ValidationException(validation.ToString());
+        return _groceryListRepository.Update(groceryList);
     }
 }
